Simon Hicks
A Minuteman from Rehoboth, Massachusetts. 

Hicks mobilised with his unit and helped seal off a British garrison in Boston after the Battles of Lexington and Concorde .

He served several short enlistments and fought in the Battle of Bennington, August 16, 1777.

After the war Hicks became something of a local celebrity and lived out his final years in in Sunderland, Vermont. He was the last person alive to have seen the Battle of Bennington.